crypto_prob_trader_plain/tradingapp/requirements.txt

python-binance>=1.0.17
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
pandas-ta>=0.3.14b0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
